Conflicts:
	bin/vendors.sh

+++ bin/build.sh
@@ -1,5 +1,12 @@
 #!/bin/sh
 
+# This file is part of the Symfony Standard Edition.
+#
+# (c) Fabien Potencier <fabien@symfony.com>
+#
+# For the full copyright and license information, please view the LICENSE
+# file that was distributed with this source code.
+
 DIR=`php -r "echo realpath(dirname(\\$_SERVER['argv'][0]));"`
 cd $DIR
 VERSION=`cat VERSION`
@@ -14,11 +21,19 @@ $DIR/app/console assets:install web/
 # Without vendors
 rm -rf /tmp/Symfony
 mkdir /tmp/Symfony
+cp -r app /tmp/Symfony/
+cp -r bin /tmp/Symfony/
+cp -r src /tmp/Symfony/
+cp -r web /tmp/Symfony/
+cp -r README.md /tmp/Symfony/
+cp -r LICENSE /tmp/Symfony/
+cp -r VERSION /tmp/Symfony/
 cd /tmp/Symfony
-git clone --depth 1 --branch master http://github.com/symfony/symfony-standard.git .
-cp $DIR/app/bootstrap* app/
-cp -r $DIR/web/bundles web/
+sudo rm -rf app/cache/* app/logs/* .git*
 chmod 777 app/cache app/logs
+
+# DS_Store cleanup
+find . -name .DS_Store | xargs rm -rf -
 
 cd ..
 # avoid the creation of ._* files
@@ -29,45 +44,63 @@ sudo rm -f $DIR/build/Symfony_Standard_$VERSION.zip
 zip -rq $DIR/build/Symfony_Standard_$VERSION.zip Symfony
 
 # With vendors
-TARGET=/tmp/Symfony/vendor
+cd $DIR
+rm -rf /tmp/vendor
+mkdir /tmp/vendor
+TARGET=/tmp/vendor
+cd $TARGET
 
-cp -r $DIR/vendor/* $TARGET/
+if [ ! -d "$DIR/vendor" ]; then
+    echo "The master vendor directory does not exist"
+    exit
+fi
+
+cp -r $DIR/vendor/* .
 
 # Assetic
-cd $TARGET
 cd assetic && rm -rf phpunit.xml* README* tests
+cd $TARGET
 
 # Doctrine ORM
+cd doctrine && rm -rf UPGRADE* build* bin tests tools lib/vendor
 cd $TARGET
-cd doctrine && rm -rf UPGRADE* build* bin tests tools lib/vendor/Symfony lib/vendor/doctrine-dbal/lib/vendor
+
+# Doctrine DBAL
+cd doctrine-dbal && rm -rf bin build* tests lib/vendor
+cd $TARGET
+
+# Doctrine Common
+cd doctrine-common && rm -rf build* tests lib/vendor
+cd $TARGET
 
 # Swiftmailer
-cd $TARGET
 cd swiftmailer && rm -rf CHANGES README* build* docs notes test-suite tests create_pear_package.php package*
+cd $TARGET
 
 # Symfony
-cd $TARGET
 cd symfony && rm -rf README.md phpunit.xml* tests *.sh vendor
+cd $TARGET
 
 # Twig
-cd $TARGET
 cd twig && rm -rf AUTHORS CHANGELOG README.markdown bin doc package.xml.tpl phpunit.xml* test
+cd $TARGET
 
 # Twig Extensions
-cd $TARGET
 cd twig-extensions && rm -rf README doc phpunit.xml* test
+cd $TARGET
 
 # Monolog
-cd $TARGET
 cd monolog && rm -rf README.markdown phpunit.xml* tests
+cd $TARGET
 
 # cleanup
-find /tmp/Symfony -name .git | xargs rm -rf -
-find /tmp/Symfony -name .gitignore | xargs rm -rf -
-find /tmp/Symfony -name .gitmodules | xargs rm -rf -
-find /tmp/Symfony -name .svn | xargs rm -rf -
+find . -name .git | xargs rm -rf -
+find . -name .gitignore | xargs rm -rf -
+find . -name .gitmodules | xargs rm -rf -
+find . -name .svn | xargs rm -rf -
 
 cd /tmp/
+mv /tmp/vendor /tmp/Symfony/
 tar zcpf $DIR/build/Symfony_Standard_Vendors_$VERSION.tgz Symfony
 sudo rm -f $DIR/build/Symfony_Standard_Vendors_$VERSION.zip
 zip -rq $DIR/build/Symfony_Standard_Vendors_$VERSION.zip Symfony
